requirements.txt
flask>=3.0,<4.0
flask-socketio>=5.3,<6.0
python-dotenv>=1.0,<2.0
opencv-python-headless>=4.9,<5.0
ultralytics>=8.2,<9.0
scikit-learn>=1.4,<2.0
pandas>=2.2,<3.0
numpy>=1.26,<2.0
joblib>=1.3,<2.0
eventlet>=0.35,<1.0
